[Refactor] Refactor CCE node create
